fix: correct sign-in error assignment flow
Sign in result will not have multiple errors assigned at once, so an
`else if` chain is more appropriate. More crucial, the final `else` that
was meant to be a fall back was only applied to the last `if`. This
meant for anything other than the final `if` conditional, it was adding
the fallback error as well.