feat(seed-tool): enhance configuration handling for db seeding
If no connection string is provided, the tool tries to read it from
`appsettings.json`. Because of the project reference this seems to mean
that the referenced project's `appsettings.json` is the one copied on
build. This is actually ideal  so that I don't have to duplicate
the configuration, but documentation confirming this behavior was
sparse.