fix(seed-tool): fix staff SQL projection
The pool for each staffmembers correspondents is every other
staffmember. To avoid projecting this query each time, the staff were
queried once and saved as a dictionary that could be filtered.

However, the projection can't use the in-memory dictionary, as it
attempts to translate the projection to SQL, causing an error.

To fix this, the PotentialCorrespondentInfos is initialized as an empty
list in the projection, and then populated in-memory after the query
completes.

In the future, it might be better to pass the full staff list, which can
then be filtered as needed in the composer. The numbers are not really
large enough for this to cause performance issues though.