refactor: enable NRTs for Staff services
Enable nullable reference types for Staff related services and their
usage in controllers.

- The StaffFormView inputs should all be nullable, but the services
using it as an input expected non-nullable values. We now convert the
form to new Staff and Accoung DTOs with extension methods that use a new
Guard class to validate the form inputs as not null.
- Use staffDTO in StaffService
- Use staffDTO and accountDTO in StaffRegistrationService
- Update queries that might default to have nullable return values
- Update StaffInfoViewModel to expect a nullable MessagingLinkId

- Update StaffController to handle potential null returns now in
services