refactor: enable NRTs for Patient services
Enable nullable reference types for Patient related services and their
usage in controllers.

- The PatientFormView inputs should all be nullable, but the services
using it as an input expected non-nullable values. We now convert the
form to new Patient and Address DTOs with extension methods that use a
new Guard class to validate the form inputs as not null.
- Use patientDTO  in PatientService
- Update queries that might default to have nullable return values
- Update PatientInfoViewModel and PatientViewService to expect a
nullable MessagingLinkId

- Update PatientsController to handle potential null returns now in
services
- Add staffId null check for staffId user claim