refactor: enable NRTs for Login services
Enable nullable reference types for Login and DemoLogin related
services and their usage in controllers.

- The LoginStaff form inputs should all be nullable, but the service
using it as input expected non-nullable values. We now convert the form
to a new LoginStaffDTO with an extension method that uses the Guard
class to validate inputs as not null.
- Use LoginStaffDTO in LoginController

- Update DemoLoginService.LoginStaffById to return a bool success
indicator and handle null user lookups
- Update DemoLoginController to handle the bool return from
LoginStaffById