Revert "refactor: introduce CSS var based color system"
This reverts commit ad1bc3319425e5c0fe0f383ddf8cd9c8714e2f6f.

After working on implementing the previous var based system, a new
approach proved to be more efficient. This one resulted in a var being
set for every variation of every hue (e.g. --primary, --primary-dimmed,
--primary-muted, --primary-light, etc.). The new approach will allow for
dynamically adjusting the main color with classes. But so much was
changed that this commit was essentially rewritten.