fix(ui): change CSS layout to prevent display gap
The former layout would cause a gap to appear at the bottom of the page
when the content shrank to the smallest breakpoint. Switching the body
to a vertical flex layout, not only fixes the gap issue, but is more
resilient to future layout changes.

The scroll is now applied to the main container instead of the main
element within it. It might be wise to remove the wrapping
main-container div.