fix: adjust CSS to ensure sort indicators rotate
The `--sort-dir` was only defined on the sort-invert class, so while the
base svg would select the proper value as a default,
(`var(--invert-dir, 0deg)`), the hover state used
`calc(180deg - var(--sort-dir))`. This calc fails when `--sort-dir` is
not defined, resulting in an invalid value and preventing the rotation
from working.

The change sets a default value of `0deg` for `--sort-dir` on the svg,
and then the `.sort-invert` class simply overrides that value to
`180deg`. Another solution would have been to have the var in calc also
fall back to 0deg, but this is more readable and straightforward.