fix: smooth extendable button animations
The use of padding was preventing smooth animation of the button
extension. This instead uses a pseudo element to create the space
between the icon and the text.

Remove a former `nbsp;` that was creating space.

Also make sure that the non-extending part is wrapped in a span for
clarity and if any future styling is needed.