fix: stop mutating route dict from breaking links
Change the queryDict assignment to a function that returns a new dict on
 each call.

When an `asp-route-*` comes after `asp-all-route-data`, it apparently
doesn't just modify the output. It was mutating the `queryDict` assigned
to it with the values.

When passed separately as a new dict to the pagination partial, this
didn't matter as only the `currentPage` value was changed (as desired).
In the display however, the results per page dropdown makes a link for
each different value (apparently mutating the query dict each time).
When this same dict was then used to create the sort links, the results
per page was not what was passed in the model, but the last value in the
 dropdown list, causing all sort links to change to 100 results per
 page.